Trait actix_web_httpauth::extractors::AuthExtractor

pub trait AuthExtractor: Sized {
    type Error: Into<Error>;
    type Future: Future<Output = Result<Self, Self::Error>>;
    fn from_service_request(req: &ServiceRequest) -> Self::Future;
}

Trait implemented by types that can extract HTTP authentication scheme credentials from the request.

It is very similar to actix’ FromRequest trait, except it operates with a ServiceRequest struct instead, therefore it can be used in the middlewares.

You will not need it unless you want to implement your own authentication scheme.

Associated Types

type Error: Into<Error>

The associated error which can be returned.

type Future: Future<Output = Result<Self, Self::Error>>

Future that resolves into extracted credentials type.

Required methods

fn from_service_request(req: &ServiceRequest) -> Self::Future

Parse the authentication credentials from the actix’ ServiceRequest.