Struct actix_web_httpauth::middleware::HttpAuthentication

pub struct HttpAuthentication<T, F> where
    T: AuthExtractor,  { /* fields omitted */ }

Middleware for checking HTTP authentication.

If there is no Authorization header in the request, this middleware returns an error immediately, without calling the F callback.

Otherwise, it will pass both the request and the parsed credentials into it. In case of successful validation F callback is required to return the ServiceRequest back.

Methods

impl<T, F, O> HttpAuthentication<T, F> where
T: AuthExtractor,
F: Fn(ServiceRequest, T) -> O,
O: Future<Output = Result<ServiceRequest, Error>>,

pub fn with_fn(process_fn: F) -> HttpAuthentication<T, F>

Construct HttpAuthentication middleware with the provided auth extractor T and validation callback F.

impl<F, O> HttpAuthentication<BasicAuth, F> where
F: Fn(ServiceRequest, BasicAuth) -> O,
O: Future<Output = Result<ServiceRequest, Error>>,

pub fn basic(process_fn: F) -> Self

Construct HttpAuthentication middleware for the HTTP "Basic" authentication scheme.

Example

// In this example validator returns immediately,
// but since it is required to return anything
// that implements `IntoFuture` trait,
// it can be extended to query database
// or to do something else in a async manner.
async fn validator(
    req: ServiceRequest,
    credentials: BasicAuth,
) -> Result<ServiceRequest, Error> {
    // All users are great and more than welcome!
    Ok(req)
}

let middleware = HttpAuthentication::basic(validator);

impl<F, O> HttpAuthentication<BearerAuth, F> where
F: Fn(ServiceRequest, BearerAuth) -> O,
O: Future<Output = Result<ServiceRequest, Error>>,

pub fn bearer(process_fn: F) -> Self

Construct HttpAuthentication middleware for the HTTP "Bearer" authentication scheme.

Example

async fn validator(req: ServiceRequest, credentials: BearerAuth) -> Result<ServiceRequest, Error> {
    if credentials.token() == "mF_9.B5f-4.1JqM" {
        Ok(req)
    } else {
        let config = req.app_data::<Config>()
            .map(|data| data.get_ref().clone())
            .unwrap_or_else(Default::default)
            .scope("urn:example:channel=HBO&urn:example:rating=G,PG-13");

        Err(AuthenticationError::from(config).into())
    }
}

let middleware = HttpAuthentication::bearer(validator);